장바구니 API 수정

+++ routes/goods.js
@@ -3,6 +3,28 @@ const router = express.Router();
 
 const Goods = require("../schemas/goods.js");
 const Cart = require("../schemas/cart.js");
+const authMiddleware = require("../middlewares/auth-middleware.js");
+
+// 장바구니 조회 API
+router.get("/goods/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
+
+  const carts = await Cart.find({ userId }).exec();
+  const goodsIds = carts.map((cart) => cart.goodsId);
+
+  const goods = await Goods.find({ goodsId: goodsIds });
+
+  const results = carts.map((cart) => {
+    return {
+      quantity: cart.quantity,
+      goods: goods.find((item) => item.goodsId === cart.goodsId),
+    };
+  });
+
+  res.json({
+    carts: results,
+  });
+});
 
 // 상품 목록 조회 API
 router.get("/goods", async (req, res) => {
@@ -46,11 +68,14 @@ router.get("/goods/:goodsId", async (req, res) => {
   res.json({ goods: result });
 });
 
-router.post("/goods/:goodsId/cart", async (req, res) => {
+// 장바구니 등록 API
+router.post("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
   const { quantity } = req.body;
 
-  const existsCarts = await Cart.find({ goodsId });
+  // 사용자 정보(userId)를 가지고, 장바구니를 조회한다.
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
     return res.status(400).json({
       success: false,
@@ -58,31 +83,36 @@ router.post("/goods/:goodsId/cart", async (req, res) => {
     })
   }
 
-  await Cart.create({ goodsId, quantity });
+  // 해당하는 사용자의 정보(userId)를 가지고 장바구니에 상품을 등록한다.
+  await Cart.create({ userId, goodsId, quantity });
 
   res.json({ result: "success" });
 })
 
-router.put("/goods/:goodsId/cart", async (req, res) => {
+// 장바구니 수정 API
+router.put("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
   const { quantity } = req.body;
 
-  const existsCarts = await Cart.find({ goodsId });
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
     await Cart.updateOne(
-      { goodsId: goodsId },
+      { userId, goodsId: goodsId },
       { $set: { quantity: quantity } }
     )
   }
   res.status(200).json({ success: true });
 })
 
-router.delete("/goods/:goodsId/cart", async (req, res) => {
+// 장바구니 삭제 API
+router.delete("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
 
-  const existsCarts = await Cart.find({ goodsId });
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
-    await Cart.deleteOne({ goodsId });
+    await Cart.deleteOne({ userId, goodsId });
   }
 
   res.json({ result: "success" });
